[docker] Add ARGs for Gazebo dependencies.
Support Gazebo 7, 8, and 9 source install utilizing a more generic
Dockerfile as template by adding ARGs that are passed when docker
build is called.

We are now using the new official .yml files to keep track of
the source repos of the dependencies for each corresponding Gazebo
version. These .yml config files are used by the vcstool and then
the downloaded repos are built using the colcon tool.